Move a bunch of regression tests from FbxObject to Base.
Now we try every FbxObject subclass for creating with a null manager, or with a
null name, or with a zombie manager, or just creating it in normal circumstances.

Generalized Base to know what type it's operating on (which must be an
FbxObject or subclass). Now we don't need to write Create() for every subclass.