Temporarily FbxProperty is not a weak-pointer. Fix this monday, etc etc.
This means if we hold a property, destroy its object, then try to use the
property, we crash.

The reason to disable it is that some properties (e.g. LclTranslation) are
member variables, so the weak-pointer mechanism doesn't notice they're being
freed when the parent object (e.g. FbxNode) is freed. In ship we wouldn't
really notice except sometimes we'd crash. In debug, the memory debugging code
notices the problem right away and asserts.