Added FbxDataTypes, and lots of support for that.
Added a script to find all the constants in the file (there's over a hundred).

Generalized the running of python scripts in the cmakelists, so it's easier to add.

Added functions in FbxProperty that use the FbxDataTypes.

This doesn't include exposing EFbxType; I want to merge in Viktoria's work for that.