Devirtualize when using generics.
When we call x.foo() and x is of generic type, the CIL helpfully says what the
type by emitting a "constrained" instruction just before the "callvirt"
instruction.

I use that in the coverage tester to determine what I'm actually calling when
calling a virtual function.

The next step is to make the virtual call using a generic function, to force
the compiler to emit that CIL. That's what the DisposeTester does.