Update tests for FbxLight.
Also track down what's going on with the Color property. There's one in
NodeAttribute and one in Light. They end up mapping to the same underlying
property handle, so there's no shotgun pointed at users' feet -- just at FBX
SDK developers' feet. I left in an assert in the test to make sure that we
catch it if/when things change.