UNI-19090: gizmos to help Joel compare skeletons
New menu command FbxSharp -> Add Comparison Gizmos (and remove).

That adds gizmos to the transforms in that hierarchy.

When you select a node in the hierarchy, it and its children show up with
red/green/blue gizmos (x/y/z) and with a black line to the parent.

Gizmo size scales down by default, so leaf nodes have smaller gizmos than
parents. You can change the size of a gizmo to affect the nodes below it in its
hierarchy.

Gizmo color you can scale up/down to help keep track which node is in which
hierarchy. Scaling can go channel by channel.