Allow building with Unity 2018.1
When we look for mono on OSX we now look for it both in the Monodevelop.app
(where it used to be) but also in Unity.app (where is now is, as of 2018.1).

While I was at it I bulletproofed the code to find Unity and to find the
assemblies.

Added dotnet_find_library to handle the looking for libraries that end in
.dll on all platforms (though it doesn't seem to work if you don't suffix
with .dll anyways).

Target for this is to copy these files to the Alembic github repo.