Fixes for Mac and Linux
Mac and Linux SWIG were compiled and have hard coded paths to swig.swg.
Set the correct location in the environment for the build, since the
configure is able to set it for itself.

Also revert the CMake version bump because Ubuntu Bionic only has CMake
v3.10.2 (and it works fine without CMP0074)